Fix generics for CustomAggregatingEnricher sensors

+++ software/webapp/src/main/java/brooklyn/entity/webapp/DynamicWebAppCluster.java
@@ -6,7 +6,9 @@ import brooklyn.entity.proxying.ImplementedBy;
 import brooklyn.event.AttributeSensor;
 
 /**
- * DynamicWebAppClusters provide cluster-wide aggregates of entity attributes.  Currently totals and averages:
+ * DynamicWebAppClusters provide cluster-wide aggregates of entity attributes.
+ * <p>
+ * Currently totals and averages:
  * <ul>
  *   <li>Entity request counts</li>
  *   <li>Entity error counts</li>
@@ -24,12 +26,12 @@ public interface DynamicWebAppCluster extends DynamicCluster, WebAppService {
     /** @deprecated since 0.5.0, just use REQUEST_COUNT_PER_NODE */
     AttributeSensor<Double> AVERAGE_REQUEST_COUNT = REQUEST_COUNT_PER_NODE;
 
-    AttributeSensor<Integer> ERROR_COUNT_PER_NODE = Attributes.newAttributeSensor("webapp.reqs.errors.perNode", "Cluster entity request error average");
+    AttributeSensor<Double> ERROR_COUNT_PER_NODE = Attributes.newAttributeSensor("webapp.reqs.errors.perNode", "Cluster entity request error average");
 
     /** @deprecated since 0.5.0, just use WebAppServiceConstants.ERROR_COUNT */
     AttributeSensor<Integer> TOTAL_ERROR_COUNT = WebAppServiceConstants.ERROR_COUNT;
     /** @deprecated since 0.5.0, use ERROR_COUNT_PER_NODE */
-    AttributeSensor<Integer> AVERAGE_ERROR_COUNT = ERROR_COUNT_PER_NODE;
+    AttributeSensor<Double> AVERAGE_ERROR_COUNT = ERROR_COUNT_PER_NODE;
 
     AttributeSensor<Double> REQUESTS_PER_SECOND_LAST_PER_NODE = Attributes.newAttributeSensor("webapp.reqs.perSec.last.perNode", "Reqs/sec (last datapoint) averaged over all nodes");
